Added checks
Added checks. Will need to add `check` package somewhere in the
socialize ecosystem. Checks are needed for applications to run properly
if the package `audit-argument-checks` is included. It is a good
practice and increases security of the app.

Another option is to remove publications completely like in the other
packages.